pyproject.toml
[project]
name = "generalist-generative-agent"
version = "0.1.0"
description = "Agentic architectural design exploration with metaphor, interpretation, modelling, and evaluation agents."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "langchain-core==0.2.27",
    "langchain-openai==0.1.20",
    "pandas==2.2.2",
    "Pillow==10.4.0",
    "python-dotenv==1.0.1",
    "rich==13.7.1",
]

[tool.uv]
package = false
